Работа с БД, обновление игры

+++ handlers/users/new_add.py
@@ -2,11 +2,12 @@ from aiogram import types
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters import Text
 
-from keyboards.inline.choose_game_kb import game_kb
+from keyboards.inline.choose_game_kb import game_kb, get_edit_game_kb
+from keyboards.inline.edit_add import edit_kb_add
 from keyboards.inline.edit_self_keyboard import edit_kb
 from loader import dp
 from states.get_info import GetInfo
-from utils.db_api.db_commands import get_user_info, get_game_by_id, get_user, create_row_adds
+from utils.db_api.db_commands import get_user_info, get_game_by_id, get_user, create_row_adds, get_info_row, update_game
 
 
 @dp.message_handler(text='📩 Новое объявление')
@@ -18,10 +19,18 @@ async def new_add_user(message: types.Message):
     if data:
         await message.answer('❌ У тебя не заполнена личная информация.\n', reply_markup=edit_kb)
     else:
-        keyboard = await game_kb()
-        await message.delete()
+        add = await get_info_row(message.from_user.id)
 
-        await message.answer('Выбери игру из списка ниже:', reply_markup=keyboard)
+        if add is not None:
+            await message.answer('Ваше объявление:\n\n'
+                                 f'{add.name} ({add.nickname}), {add.age} лет, {add.game}\n\n'
+                                 f'{add.description}\n\n'
+                                 f'Выберите то, что хотите изменить:', reply_markup=edit_kb_add)
+        else:
+            keyboard = await game_kb()
+            await message.delete()
+
+            await message.answer('Выбери игру из списка ниже:', reply_markup=keyboard)
 
 
 @dp.callback_query_handler(Text(startswith='game_'))
@@ -65,7 +74,26 @@ async def get_description(message: types.Message, state: FSMContext):
         name=user_data.name,
         nickname=user_data.nickname,
         age=user_data.age,
-        description=description
+        description=description,
+        game=game
     )
 
     await state.reset_state(True)
+
+
+@dp.callback_query_handler(Text(equals='edit_game'))
+async def edit_user_game(call: types.CallbackQuery):
+    keyboard = await get_edit_game_kb()
+
+    await call.message.edit_text('Выбери игру из списка ниже:', reply_markup=keyboard)
+
+
+@dp.callback_query_handler(Text(startswith='egame_'))
+async def get_edit_game(call: types.CallbackQuery):
+    data = int(call.data.split('_')[1])
+    game = await get_game_by_id(data)
+    await update_game(
+        owner_id=call.from_user.id,
+        game=game.name
+    )
+    await call.message.edit_text(f'Вы успешно изменили игру на {game.name}')
